mod: #16 API 변경 사항 반영

+++ lib/data/models/analysis_item_create_request.dart
@@ -7,7 +7,7 @@ class AnalysisItemCreateRequest {
   final String contactType;
   final String sourceUrl;
   final String notes;
-  final int salary;
+  final double salary;
 
   const AnalysisItemCreateRequest({
     required this.imageNames,
